+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Translations: [简体中文](CHANGELOG_zh.md)
 > 2. The maven groupId is upgraded to `io.github.panpf.sketch4`, so versions 2.\* and 3.\* will not
      prompt for upgrade.
 
+# new
+
+* fix: Fixed a bug where SkiaAnimatedImagePainter crashed when encountering disguised gif
+  images. [#205](https://github.com/panpf/sketch/issues/205)
+
 # 4.0.0-alpha05
 
 * fix: Fix the bug of exception when AppIconUriFetcher.Factory parses